+++ src/onegov.ballot/onegov/ballot/__init__.py
@@ -8,6 +8,7 @@ from onegov.ballot.models import Candidate
 from onegov.ballot.models import CandidateResult
 from onegov.ballot.models import ComplexVote
 from onegov.ballot.models import Election
+from onegov.ballot.models import ElectionAssociation
 from onegov.ballot.models import ElectionCompound
 from onegov.ballot.models import ElectionResult
 from onegov.ballot.models import List
@@ -26,6 +27,7 @@ __all__ = [
     'CandidateResult',
     'ComplexVote',
     'Election',
+    'ElectionAssociation',
     'ElectionCollection',
     'ElectionCompound',
     'ElectionCompoundCollection',