+++ src/onegov.ballot/onegov/ballot/models/__init__.py
@@ -10,6 +10,7 @@ See:
 from onegov.ballot.models.election import Candidate
 from onegov.ballot.models.election import CandidateResult
 from onegov.ballot.models.election import Election
+from onegov.ballot.models.election import ElectionAssociation
 from onegov.ballot.models.election import ElectionCompound
 from onegov.ballot.models.election import ElectionResult
 from onegov.ballot.models.election import List
@@ -30,6 +31,7 @@ __all__ = [
     'CandidateResult',
     'ComplexVote',
     'Election',
+    'ElectionAssociation',
     'ElectionCompound',
     'ElectionResult',
     'List',