+++ src/onegov.ballot/HISTORY.rst
@@ -4,6 +4,9 @@ Changelog
 - Adds a majority type attribute to elections.
   [msom]
 
+- Allows to store related elections.
+  [msom]
+
 3.5.0 (2018-04-26)
 ~~~~~~~~~~~~~~~~~~~